// src/store/modules/user/action-type.ts
/*
 * @Description: 
 * @Author: yangzai
 * @Date: 2021-05-20 10:38:38
 * @LastEditTime: 2021-05-20 10:43:14
 * @LastEditors: yangzai
 */
export enum UserActionTypes {
    ACTION_LOGIN = 'ACTION_LOGON',
    ACTION_RESET_TOKEN = 'ACTION_RESET_TOKEN',
    ACTION_GET_USER_INFO = 'ACTION_GET_USER_INFO',
    ACTION_CHANGE_ROLES = 'ACTION_CHANGE_ROLES',
    ACTION_LOGIN_OUT = 'ACTION_LOGIN_OUT'
}