// src/store/modules/user/mutation-type.ts
/*
 * @Description: 
 * @Author: yangzai
 * @Date: 2021-05-20 10:38:19
 * @LastEditTime: 2021-05-20 10:46:51
 * @LastEditors: yangzai
 */
export enum UserMutationTypes {
    SET_TOKEN = 'SET_TOKEN',
    SET_NAME = 'SET_NAME',
    SET_AVATAR = 'SET_AVATAR',
    SET_INTRODUCTION = 'SET_INTRODUCTION',
    SET_ROLES = 'SET_ROLES',
    SET_EMAIL = 'SET_EMAIL'
}